requirements.txt
pandas==2.2.0
requests>=2.25.0
chardet>=4.0.0
openpyxl>=3.0.0
pysftp>=0.2.9
python-dotenv>=0.19.0